重构t_baseline表

+++ cat-home/src/main/java/com/dianping/cat/core/mybatis/baseline/dao/BaselineMapper.java
@@ -7,9 +7,9 @@ import org.apache.ibatis.annotations.Param;
 import com.dianping.cat.core.mybatis.baseline.dao.data.BaselineDO;
 
 public interface BaselineMapper {
-	int deleteByPrimaryKey(@Param("id") Integer id);
+	int deleteByPrimaryKey(@Param("id") Long id);
 
-	BaselineDO findByPrimaryKey(@Param("id") Integer id);
+	BaselineDO findByPrimaryKey(@Param("id") Long id);
 
 	int insert(BaselineDO record);
